Add Cordova example to use new Cordova plugin
The best practices of building Cordova apps is not commit the platforms/ or plugins/ directories, since those should be able to be blown away and rebuilt by Cordova at any time.

The JS has been updated to reflect that the `Realm` constructor is not available until the 'deviceready' event has fired. It is completely expected in Cordova apps that plugins will not have initialized until that time.